refactor: Consolidate dfx-core identity errors into error/identity.rs
