--- src/dfx-core/src/error/identity/convert_mnemonic_to_key.rs
@@ -1,7 +1,0 @@
-use thiserror::Error;
-
-#[derive(Error, Debug)]
-pub enum ConvertMnemonicToKeyError {
-    #[error("Failed to derive extended secret key from path")]
-    DeriveExtendedKeyFromPathFailed(#[source] bip32::Error),
-}
