--- src/dfx-core/src/error/identity/create_identity_config.rs
@@ -1,8 +1,0 @@
-use crate::error::encryption::EncryptionError;
-use thiserror::Error;
-
-#[derive(Error, Debug)]
-pub enum CreateIdentityConfigError {
-    #[error("Failed to generate a fresh encryption configuration")]
-    GenerateFreshEncryptionConfigurationFailed(#[source] EncryptionError),
-}
