--- src/dfx-core/src/error/identity/get_identity_config_or_default.rs
@@ -1,8 +1,0 @@
-use crate::error::structured_file::StructuredFileError;
-use thiserror::Error;
-
-#[derive(Error, Debug)]
-pub enum GetIdentityConfigOrDefaultError {
-    #[error("Failed to load configuration for identity '{0}'")]
-    LoadIdentityConfigurationFailed(String, #[source] StructuredFileError),
-}
