--- src/dfx-core/src/error/identity/get_legacy_credentials_pem_path.rs
@@ -1,8 +1,0 @@
-use crate::error::get_user_home::GetUserHomeError;
-use thiserror::Error;
-
-#[derive(Error, Debug)]
-pub enum GetLegacyCredentialsPemPathError {
-    #[error("Failed to get legacy pem path")]
-    GetLegacyPemPathFailed(#[source] GetUserHomeError),
-}
