--- src/dfx-core/src/error/identity/new_hardware_identity.rs
@@ -1,8 +1,0 @@
-use ic_identity_hsm::HardwareIdentityError;
-use thiserror::Error;
-
-#[derive(Error, Debug)]
-pub enum NewHardwareIdentityError {
-    #[error("Failed to instantiate hardware identity for identity '{0}'")]
-    InstantiateHardwareIdentityFailed(String, #[source] Box<HardwareIdentityError>),
-}
