--- src/dfx-core/src/error/identity/write_default_identity.rs
@@ -1,8 +1,0 @@
-use crate::error::structured_file::StructuredFileError;
-use thiserror::Error;
-
-#[derive(Error, Debug)]
-pub enum WriteDefaultIdentityError {
-    #[error("Failed to save identity manager configuration")]
-    SaveIdentityManagerConfigurationFailed(#[source] StructuredFileError),
-}
